Pull score_matches and time_boxed into a separate module
